Hide links before onload.
When opening the Candy page in mobile mode, the links would remain exposed at the top of the page. This is because the videos are loaded before window.onload is called. This is a small hack to prevent the dropdown from showing up at the start. There should, however, be a better solution to this and the video loading problem.